+++ qnexus/client/hugr.py
@@ -27,12 +27,11 @@ from qnexus.models import QuantinuumConfig
 from qnexus.models.annotations import Annotations, CreateAnnotations, PropertiesDict
 from qnexus.models.filters import (
     CreatorFilter,
-    FuzzyNameFilter,
+    NameFilter,
     PaginationFilter,
     ProjectRefFilter,
     PropertiesFilter,
     ScopeFilter,
-    ScopeFilterEnum,
     SortFilter,
     SortFilterEnum,
     TimeFilter,
@@ -43,12 +42,13 @@ from qnexus.models.references import (
     HUGRRef,
     ProjectRef,
 )
+from qnexus.models.scope import ScopeFilterEnum
 
 
 class Params(
     SortFilter,
     PaginationFilter,
-    FuzzyNameFilter,
+    NameFilter,
     CreatorFilter,
     ProjectRefFilter,
     PropertiesFilter,
@@ -61,7 +61,9 @@ class Params(
 @merge_scope_from_context
 @merge_project_from_context
 def get_all(
+    *,
     name_like: str | None = None,
+    name_exact: list[str] | None = None,
     creator_email: list[str] | None = None,
     project: ProjectRef | None = None,
     properties: PropertiesDict | None = None,
@@ -78,6 +80,7 @@ def get_all(
 
     params = Params(
         name_like=name_like,
+        name_exact=name_exact,
         creator_email=creator_email,
         properties=properties,
         project=project,
@@ -131,6 +134,7 @@ def _to_hugr_ref(page_json: dict[str, Any]) -> DataframableList[HUGRRef]:
 def get(
     *,
     id: Union[UUID, str, None] = None,
+    name: str | None = None,
     name_like: str | None = None,
     creator_email: list[str] | None = None,
     project: ProjectRef | None = None,
@@ -154,6 +158,7 @@ def get(
 
     return get_all(
         name_like=name_like,
+        name_exact=[name] if name else None,
         creator_email=creator_email,
         properties=properties,
         project=project,
@@ -338,11 +343,6 @@ def _fetch_hugr_package(
     """Utility method for fetching a HUGR Package from a HUGRRef."""
 
     hugr_bytes = _fetch_hugr_bytes(handle=handle, scope=scope)
-
-    raise qnx_exc.ResourceFetchFailed(
-        message="Converting to HUGR Package is currently unavailable.",
-        status_code=400,
-    )
     return Package.from_bytes(envelope=hugr_bytes)