+++ qnexus/client/projects.py
@@ -17,16 +17,16 @@ from qnexus.models.annotations import Annotations, CreateAnnotations, Properties
 from qnexus.models.filters import (  # PropertiesFilter, # Not yet implemented
     ArchivedFilter,
     CreatorFilter,
-    FuzzyNameFilter,
+    NameFilter,
     PaginationFilter,
     ScopeFilter,
-    ScopeFilterEnum,
     SortFilter,
     SortFilterEnum,
     TimeFilter,
 )
 from qnexus.models.job_status import WAITING_STATUS, JobStatusEnum
 from qnexus.models.references import DataframableList, ProjectRef
+from qnexus.models.scope import ScopeFilterEnum
 
 # Colour-blind friendly colours from https://www.nature.com/articles/nmeth.1618
 _COLOURS = ["#e69f00", "#56b4e9", "#009e73", "#f0e442", "#0072b2", "#d55e00", "#cc79a7"]
@@ -36,7 +36,7 @@ class Params(
     ScopeFilter,
     SortFilter,
     PaginationFilter,
-    FuzzyNameFilter,
+    NameFilter,
     CreatorFilter,
     # PropertiesFilter, # Not yet implemented
     TimeFilter,
@@ -47,7 +47,9 @@ class Params(
 
 @merge_scope_from_context
 def get_all(
+    *,
     name_like: str | None = None,
+    name_exact: list[str] | None = None,
     creator_email: list[str] | None = None,
     created_before: datetime | None = None,
     created_after: datetime | None = datetime(day=1, month=1, year=2023),
@@ -63,6 +65,7 @@ def get_all(
 
     params = Params(
         name_like=name_like,
+        name_exact=name_exact,
         creator_email=creator_email,
         created_before=created_before,
         created_after=created_after,
@@ -106,6 +109,7 @@ def _to_projectref(data: dict[str, Any]) -> DataframableList[ProjectRef]:
 def get(
     *,
     id: Union[str, UUID, None] = None,
+    name: str | None = None,
     name_like: str | None = None,
     creator_email: list[str] | None = None,
     created_before: datetime | None = None,
@@ -127,6 +131,7 @@ def get(
 
     return get_all(
         name_like=name_like,
+        name_exact=[name] if name else None,
         creator_email=creator_email,
         created_before=created_before,
         created_after=created_after,
@@ -153,7 +158,7 @@ def get_or_create(
         properties=properties,
     )
     try:
-        return get(name_like=annotations.name)
+        return get(name=annotations.name)
     except qnx_exc.ZeroMatches:
         return create(
             name=annotations.name,