+++ qnexus/client/teams.py
@@ -3,8 +3,8 @@
 import qnexus.exceptions as qnx_exc
 from qnexus.client import get_nexus_client
 from qnexus.context import merge_scope_from_context
-from qnexus.models.filters import ScopeFilterEnum
 from qnexus.models.references import DataframableList, TeamRef
+from qnexus.models.scope import ScopeFilterEnum
 
 
 @merge_scope_from_context