+++ qnexus/client/users.py
@@ -5,8 +5,8 @@ from uuid import UUID
 import qnexus.exceptions as qnx_exc
 from qnexus.client import get_nexus_client
 from qnexus.context import merge_scope_from_context
-from qnexus.models.filters import ScopeFilterEnum
 from qnexus.models.references import UserRef
+from qnexus.models.scope import ScopeFilterEnum
 
 
 def get_self() -> UserRef: