+++ qnexus/client/wasm_modules.py
@@ -20,24 +20,24 @@ from qnexus.context import (
 from qnexus.models.annotations import Annotations, CreateAnnotations, PropertiesDict
 from qnexus.models.filters import (
     CreatorFilter,
-    FuzzyNameFilter,
+    NameFilter,
     PaginationFilter,
     ProjectRefFilter,
     PropertiesFilter,
     ScopeFilter,
-    ScopeFilterEnum,
     SortFilter,
     SortFilterEnum,
     TimeFilter,
 )
 from qnexus.models.references import DataframableList, ProjectRef, WasmModuleRef
+from qnexus.models.scope import ScopeFilterEnum
 
 
 class Params(
     ScopeFilter,
     SortFilter,
     PaginationFilter,
-    FuzzyNameFilter,
+    NameFilter,
     CreatorFilter,
     ProjectRefFilter,
     PropertiesFilter,
@@ -49,7 +49,9 @@ class Params(
 @merge_scope_from_context
 @merge_project_from_context
 def get_all(
+    *,
     name_like: str | None = None,
+    name_exact: list[str] | None = None,
     creator_email: list[str] | None = None,
     project: ProjectRef | None = None,
     properties: PropertiesDict | None = None,
@@ -66,6 +68,7 @@ def get_all(
 
     params = Params(
         name_like=name_like,
+        name_exact=name_exact,
         creator_email=creator_email,
         properties=properties,
         project=project,
@@ -119,6 +122,7 @@ def _to_wasm_module_ref(page_json: dict[str, Any]) -> DataframableList[WasmModul
 def get(
     *,
     id: Union[UUID, str, None] = None,
+    name: str | None = None,
     name_like: str | None = None,
     creator_email: list[str] | None = None,
     project: ProjectRef | None = None,
@@ -141,6 +145,7 @@ def get(
 
     return get_all(
         name_like=name_like,
+        name_exact=[name] if name else None,
         creator_email=creator_email,
         properties=properties,
         project=project,